第７講コメント修正#1
